Fixed documentation about pointing to a custom repo
The code just needed the url for the zip file. But it redundantly
used 2 settings. The repository and a url attribute.

We only need the url. The repository was just used to get the name
of the underlying folder which was unecessary anyways. Therefore,
named the repository to point to the actual url. Code also got cleaned up
in this process.

Also updated few dependencies