Fix: Not raising an error for shortIndex.json not found
When the cache tries to update for the first time, the shortIndex.json
file is not present. Therefore, it prints an error when it tries to
delete it.

But if the file is already not present, that means the job of unlinking
is already complete. So not raising an error in that condition