Update cache if older than 30 days
The previous behavior would still check if the cache were more than 30 days out of date, but all it would do was warn the user to run `tldr --update`.

The intention behind this fix is that because tldr CAN update itself, it should just try to update automatically when the cache is out of date instead of bothering the user.