fix running tests under windows
The tests used a linux path separator for referencing things on the
filesystem, except that then uses path.sep under the hood for path
splitting, which is / on *unix and \\ on windows. Instead of hard-coding
the separator, we can use path.join instead which will give us the
right seperator.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>