Address Race Conditions and Better Error Handling to fix Issue #256 (#316)
* Issue #256 Better Error Handling

There were a couple of anti patterns being used, and improper use of promises that lead to race conditions:

1. Using `process.exit()`  directly is usually not advisable as it prevents things from shutting down
nicely. It also means if someone is trying to use this as a library, which I think may be the
case for issue #256, you will exit from their code as well (i.e., error handling becomes impossible).

2. Functions that caught errors were printing them directly to STDERR rather than rethrowing them or
throwing a new error. Again, if someone is trying to use it as a library as in the issues, this will
make it impossible for them to do error handling.

3. There were some places where errors were being caught and printed to STDERR, but not causing the
the process to exit with a non-zero exit code. It's always advisable to exit with a non-zero exit code
when there is a failure.

4. There were some places where promises were not being waited for, leading to race conditions.

* Fix audit issues