Remove flaky test
The timeout of 30s was never enough, and this test
would fail in CI all the time.

Increasing the timeout does not make much sense as it will just
slow down things more. We should ideally mock the network to have
reliable unit tests, otherwise they are technically component tests.

Cache updates aren't a major functionality, so we just remove the test
to have a more robust CI pipeline.